Update "Twitter" to "X"
I updated the link text and its destination. No worries, by the way,
the old link redirects/works, and it's not like anyone calls Twitter "X"
but maybe this'll seem dated eventually. Maybe someone will read it and
think "Oh wow this was written way back in the days of Twitter!"